+++ src/main/java/net/tassia/pancake/util/config/ConfigIO.kt
@@ -1,4 +1,4 @@
-package net.tassia.pancake.config
+package net.tassia.pancake.util.config
 
 import java.io.*
 import kotlin.reflect.KMutableProperty
