+++ src/main/java/net/tassia/pancake/util/config/MutableCommenter.kt
@@ -1,4 +1,4 @@
-package net.tassia.pancake.config
+package net.tassia.pancake.util.config
 
 /**
  * A mutable [ConfigCommenter], allows for easily managing commenting.
